Extract HookUpStepperWithFormik component.
